[CWS] fix unary parsing in secl evaluator (#40419)
### What does this PR do?

Given the expression
```
"!"!=A
```
the parser was parsing the `!=` operator as a unary operator ! with a
null body, seemingly because participle doesn't not support grouping
across multiple struct fields.

This PR extracts the "Op + Unary" part of the Unary expression to a
different struct, allowing to express Unary as a simple choice, fixing
the issue.

This PR additionally simplifies the punctuation definition in the lexer.

This PR also extracts this test case to something in common with the
fuzzer, that is going to allow us to add reported cases and ensure
non-regression.

### Motivation

### Describe how you validated your changes

### Additional Notes