Allow passing arbitrary extra args to go test command (#40415)
### What does this PR do?
Allow passing arbitrary extra args to go test command.

### Motivation
I needed to pass the `-o` flag to `go test` so that test binaries are
kept (in order to debug a core dump generated in a test), which means
adding an extra (very uncommon) argument to the test.

### Describe how you validated your changes
Local testing.

### Additional Notes
Allowing passing arbitrary arguments to `go` subcommands is very useful
for local uses.