Return a non-nil error in python loader (#40417)
### What does this PR do?
Return a non-nil error in Python loader when a check could not be
loaded.

### Motivation
In that code block the error can be `nil`, so the loader would end up
returning `nil, nil`, and the scheduler tries to use the `nil` check
(since there was no error), resulting in a segfault.
Overall we should return a non-`nil` error here, even if we don't know
what the underlying rtloader error is.

### Describe how you validated your changes
Added a unit test for the change.
If you revert the change in `pkg/collector/python/loader.go`, the unit
test fails, and the scheduler would crash trying to use the `nil` check
returned.

### Additional Notes