[dynamic tests]: Print detected changes in evaluation job (#40441)
### What does this PR do?

Print the changes in the evaluation job. To make it easier to understand
why some tests are executed.

### Motivation

Trying to understand why we run that much tests on that job:
https://gitlab.ddbuild.io/DataDog/datadog-agent/-/jobs/1108089972
But when I do that locally I do not get the same result. I would like to
know what changes are detected in the job


### Describe how you validated your changes 

### Additional Notes