Do not import log package in pkg/obfuscate (#40376)
### What does this PR do?
Remove the import to pkg/util/log in pkg/obfuscate.

There were only two uses in the package, they are updated to use a
logger injected from the configuration.

### Motivation
Fix https://github.com/DataDog/datadog-agent/issues/40373.

### Describe how you validated your changes
CI

### Additional details
The only API change for the package is adding a `FullLogger` field in
the configuration struct, which takes precedence over the previous
`Logger` field when set (if we extended the `Logger` interface then it
would have been a breaking change).