Remove unused Windows main for iot-agent (#39999)
<!--
* Contributors are encouraged to read our
[CONTRIBUTING](/CONTRIBUTING.md) documentation.
* Both Contributor and Reviewer Checklists are available at
https://datadoghq.dev/datadog-agent/guidelines/contributing/#pull-requests.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's
change.
* Please fill the below sections if possible with relevant information
or links.
-->
### What does this PR do?
Remove some logic to build the iot agent on Windows and other binaries
in `agent-binaries` in omnibus.

### Motivation
The Windows "agent-binaires" package is made up of a mix of the Windows
iot-agent and dogstatsd packages, it is only used by
https://github.com/DataDog/datadog-aas-extension/blob/38350ce512c5aae56d379643fc2734bd83a77a8e/node/scripts/build-packages.sh#L43-L44,
but actually only the trace-agent and dogstatsd are used.
The Windows iot-agent build is not used for anything else, so a lot of
the logic can be removed.

### Describe how you validated your changes
<!--
Validate your changes before merge, ensuring that:
* Your PR is tested by static / unit / integrations / e2e tests
* Your PR description details which e2e tests cover your changes, if any
* The PR description contains details of how you validated your changes.
If you validated changes manually and not through automated tests, add
context on why automated tests did not fit your changes validation.

If you want additional validation by a second person, you can ask
reviewers to do it. Describe how to set up an environment for manual
tests in the PR description. Manual validation is expected to happen on
every commit before merge.

Any manual validation step should then map to an automated test. Manual
validation should not substitute automation, minus exceptions not
supported by test tooling yet.
-->

### Possible Drawbacks / Trade-offs

### Additional Notes
<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the
proposed
  version was chosen.
-->
Ideally we should stop using the iot-agent software definitely for that,
now it's only building the trace-agent, which really doesn't makes
sense, but it would be a bigger change.