EKS test for container now use AL2023 (#40358)
This is just a template. Please delete all unneeded sections and add
others if appropriate.

### What does this PR do?

Move containers test to new AL2023 nodes. It should remove some
flakiness due to a bug in the AL2 AMI

Note: The cluster name is explicitly set in the Agent configuration,
because the agent is not able to retrieve it properly with the new
AL2023 nodes.
The cluster name detection is made by doing a call to EC2 API and it is
failing with
```
EC2: GetNetworkID failed to get mac addresses: could not fetch token from IMDSv2: Put "http://169.254.169.254/latest/api/token": context deadline exceeded (Client.Timeout exceeded while awaiting headers)
```

### Motivation

### Describe how you validated your changes (if not by through tests)

### Possible Drawbacks / Trade-offs