dynisnt/ir,irgen: allow some pointers to not be resolved
This commit introduces the notion of an ir.UnresolvedPointeeType which
corresponds to a type we don't know anything about beyond its name, but
we might collect a pointer to it. Practically speaking we should never
be able to collect one of these because of pointer chasing limits in
bpf.

This new type is crucial to be able to have a well-formed type graph
without resolving everything.

This change doesn't perfectly model the minimal amount of type
resolution; it takes the maximum depth across all probes and then
resolves types to that depth.

For testing, this change now makes sure that as we increase the depth,
eventually we converge on the set of types we'd find if we have the
maximal depth we tend to use in our probes.