Reduce the time waited before checking state of services (#39291)
<!--
* Contributors are encouraged to read our
[CONTRIBUTING](/CONTRIBUTING.md) documentation.
* Both Contributor and Reviewer Checklists are available at
https://datadoghq.dev/datadog-agent/guidelines/contributing/#pull-requests.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's
change.
* Please fill the below sections if possible with relevant information
or links.
-->
### What does this PR do?

Reduce the time to wait before checking the state of services in windows
tests. It used to be an issue because some powershell process were
leaked but with the merge of:
https://github.com/DataDog/datadog-agent/pull/38667 it should no longer
be.

We can see these jobs are in the critical path, for example on a main
pipeline:
[Datadog](https://app.datadoghq.com/ci/pipeline-executions?query=ci_level%3Apipeline%20%40ci.pipeline.name%3A%22DataDog%2Fdatadog-agent%22%20%40git.branch%3Amain%20%40ci.pipeline.id%3A74716074&agg_m=count&agg_m_source=base&agg_t=count&buildId=AwAAAZjmL8S4TUjrrwAAABhBWmptTDhTNEFBQ0w0eVgxZzQ4b0FEeVgAAAAkMTE5OGU2MzItYjJiOC00MzYwLTljYjEtOWNlYjQwMzY3OTViAABGxw&fromUser=false&index=cipipeline&span_id=129955806217856033&tab=overview&trace=AwAAAZjmL8S4TUjrrwAAABhBWmptTDhTNEFBQ0w0eVgxZzQ4b0FEeVgAAAAkMTE5OGU2MzItYjJiOC00MzYwLTljYjEtOWNlYjQwMzY3OTViAABGxw&trace_filters=%5Cc%5Cr%5Ci%5Ct%5Ci%5Cc%5Ca%5Cl%5C_%5Cp%5Ca%5Ct%5Ch%5C&start=1755604465102&end=1756209265102&paused=false)

Job duration decrease from 45 minutes in average to 15 minutes

### Motivation

Make pipeline faster. With that change we can have a pipeline running
all the e2e tests completing in 90 minutes.

### Describe how you validated your changes
<!--
Validate your changes before merge, ensuring that:
* Your PR is tested by static / unit / integrations / e2e tests
* Your PR description details which e2e tests cover your changes, if any
* The PR description contains details of how you validated your changes.
If you validated changes manually and not through automated tests, add
context on why automated tests did not fit your changes validation.

If you want additional validation by a second person, you can ask
reviewers to do it. Describe how to set up an environment for manual
tests in the PR description. Manual validation is expected to happen on
every commit before merge.

Any manual validation step should then map to an automated test. Manual
validation should not substitute automation, minus exceptions not
supported by test tooling yet.
-->

### Possible Drawbacks / Trade-offs

### Additional Notes
<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the
proposed
  version was chosen.
-->