[CSPM] add real config option to enable compliance system-probe module (#40450)
### What does this PR do?

This PR adds a new `compliance_config.enabled` to enable the compliance
system-probe module. This is in replacement to
the currently used `runtime_security_config.compliance_module.enabled`
and in preparation to being able to use the compliance module without
CWS (runtime security).

This module is only used internally for now, so the transition should be
easy, and we will be able to remove the old config flag in 1 or 2
versions.

### Motivation

### Describe how you validated your changes

### Additional Notes