dyninst/symdb: don't Seek() across compile units (#40402)
The standard library has a bug (*) causing Reader.Seek() to be broken
when seeking across compilation units when iterating over DWARF v5
(which is used by binaries built with Go 1.25+). That bug is fixed in Go
1.25, but we use 1.24 at the moment. Stop doing Seek() in the symdb code
until we upgrade. This doesn't seem to have a big performance impact.

This patch is not sufficient for supporting Go 1.25, so we don't include
1.25 in test binaries yet.

* see https://go-review.googlesource.com/c/go/+/655976

[DEBUG-4385]

[DEBUG-4385]:
https://datadoghq.atlassian.net/browse/DEBUG-4385?atlOrigin=eyJpIjoiNWRkNTljNzYxNjVmNDY3MDlhMDU5Y2ZhYzA5YTRkZjUiLCJwIjoiZ2l0aHViLWNvbS1KU1cifQ