[dynamic tests]: Introduce a file based index to compare accuracy against package based index (#40424)
### What does this PR do?

Introduce and evaluate a new index that relies on code coverage, instead
of considering only package, we now consider each file in the coverage
report. The goal is to have a more precise index and to evaluate if this
extra accuracy can make us skipping some test that we should not.

### Motivation

Running only relevant test, and trying to run as few of them as possible

### Describe how you validated your changes

### Additional Notes