[CWS] reduce kernel memory usage of `pid_rate_limiters` and `inet_bind_args` (#40492)
### What does this PR do?

Those 2 maps have very low utilization, this PR reduces drastically the
max entries to limit the waste in kernel memory.

`pid_rate_limiters`: is only used to rate limit the ptrace and setrlimit
events, on production we store on average 2 values in this map that is
currently limited to ~131k entries.

`inet_bind_args`: is only used to store args between `inet_bind` entry
and exit, on production we store on average 4 values in this map that is
currently limited to ~131k entries.

This PR divides by 4 the max entries used by those 2 maps.

### Motivation

### Describe how you validated your changes

### Additional Notes