process: service: make service heartbeats do partial updates (#40429)
### What does this PR do?

This PR make the process collector, when service collection is enabled,
able to request partial Service updates from system-probe on heartbeats.
This is done by separating the `pids` parameter of the requests made to
SP between new pids and "heartbeat pids". The SP `/discovery/services`
path will now perform full collection of service info on the new pids,
as before, and collect only the ports and log files on pids that need a
heartbeat.

### Motivation

DSCVR-188
In short, with the removal of service info caching in system-probe, the
existing code was re-computing every fields on heartbeats, including
service names, language, etc... which are expensive operations and
unneeded.

This PR prevents the expensive updates and should reduce the number of
allocations made by service updates.

### Describe how you validated your changes

Change covered by unit tests.

### Additional Notes