[CWS] add support for activity dumps in direct sender (#40519)
### What does this PR do?

With the `direct_send_from_system_probe` you can already send regular
events directly from the system-probe (skipping the security agent),
this PR adds support for activity dumps as well.

To do that, some of the existing direct sender code is extracted and
repurposed to support both payload types.

This feature is disabled by default, and is part of a much broader
migration.

### Motivation

### Describe how you validated your changes

### Additional Notes