RLS-20: Enable K8s SSI gradual rollout support in DCA (#39915)
<!--
* Contributors are encouraged to read our
[CONTRIBUTING](/CONTRIBUTING.md) documentation.
* Both Contributor and Reviewer Checklists are available at
https://datadoghq.dev/datadog-agent/guidelines/contributing/#pull-requests.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's
change.
* Please fill the below sections if possible with relevant information
or links.
-->
### What does this PR do?
The goal of this PR is to enable gradual rollout support for K8s SSI
users. The breakdown of the changes is as follows:

**The `ImageResolver`**
The `ImageResolver` will be initialized in the `start` command for DCA,
and use the existing remote config client (if remote config is enabled)
to reference `K8S_INJECTION_DD`. `K8S_INJECTION_DD` is a DD managed
config that is split up by repository (`apm-inject` and
`dd-lib-[lang]-init`), and for each repo mutable tags (`latest`, `3`,
`3.1`, etc...) are mapped to image digests (see
[here](https://github.com/DataDog/rc-employee-configurations/tree/main/configs/K8S_INJECTION_DD/generated_versions)).
The `ImageResolver` will use this mapping to determine which version of
the tracer library and the injector to use (image reference with
specific image digest like
`gcr.io/datadoghq/dd-lib-python-init@sha123`). If remote config is not
enabled, it will default to the current behavior (image reference with a
mutable tag like `gcr.io/datadoghq/dd-lib-python-init:3`). DD will
control the updates and deployments of `K8S_INJECTION_DD`, and gradually
release the new mappings to additional subsets of organizations.

<img width="3529" height="3136" alt="imageresolver"
src="https://github.com/user-attachments/assets/5aa8389c-607a-44b2-a656-acea1e00ad22"
/>


**Configuring DCA to support gradual rollout**
- **TL;DR** - We need to initialize an `ImageResolver` in the `start`
command for the DCA, and pass the `ImageResolver` down to the
`MutatorCore` as well as the `language.libInfoWithResolver(...)` method.

**Mutating the pod spec**
- **TL;DR** - We need to pass the `ImageResolver` from the `MutatorCore`
to the `language.libInfoWithResolver(...)` method again. Since the
`MutatorCore` already has its `ImageResolver` at this point, it can
directly reference it when calling `MutatorCore.newInjector(...)` when
initializing a new injector instance.


### Motivation
Currently, new images are released to SSI customers all at once. This
means we risk causing significant impact on customers with faulty
releases, and we have no way to slowly monitor a release and how it does
on customer applications. This change will make the gradual rollout
possible, as well as give Datadog the ability to halt a release from
spreading if it's deemed problematic.

### Describe how you validated your changes
<!--
Validate your changes before merge, ensuring that:
* Your PR is tested by static / unit / integrations / e2e tests
* Your PR description details which e2e tests cover your changes, if any
* The PR description contains details of how you validated your changes.
If you validated changes manually and not through automated tests, add
context on why automated tests did not fit your changes validation.

If you want additional validation by a second person, you can ask
reviewers to do it. Describe how to set up an environment for manual
tests in the PR description. Manual validation is expected to happen on
every commit before merge.

Any manual validation step should then map to an automated test. Manual
validation should not substitute automation, minus exceptions not
supported by test tooling yet.
-->
- Local E2E testing via `injector-dev`
- Added new unit tests for `ImageResolver` functionality
- Updated pre-existing tests pertaining to image references for the
apm-injector and our tracer library images.

### Possible Drawbacks / Trade-offs
- One known trade-off with this gradual rollout plan is that this will
inevitably lengthen the release process for SSI artifacts. We will need
to let rollout phases happen over a non-trivial amount of time to ensure
we have enough signal to proceed with the rollout (over a few days to a
week), whereas current state it is immediate (same day). This is an
acceptable trade-off, as it provides us the confidence to have more
customers use our releases, and allows us to be more proactive instead
of reactive to possible widespread issues.

### Additional Notes
<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the
proposed
  version was chosen.
-->
- We will be monitoring injection success/error rates as this goes
along. While success rates are deemed healthy, the rollout will
continue. If at any point we have concerns, the `K8S_INJECTION_DD`
rollout will be stopped and new organizations and new pods will not be
able to pull the faulty image. Important to call out here that this is
just "stopping the bleeding" and will not proactively revert any image
versions on existing pods.

---------

Co-authored-by: Esther Kim <esther.kim@datadoghq.com>