Add pquery.bzl tool (#40408)
pquery is a tool to pretty print providers of a target. This is useful
for rule development.

Usage:
`bazel cquery --output=starlark --starlark:file=${PATH_TO_THIS_FILE}
//path/to/pkg:target`

For fun, try it:
```
bazel cquery --output=starlark --starlark:file=bazel/tools/pquery.bzl //deps:smoketest
```

I did not include a wrapper to invoke this easily because it's too
early. We could potentially share this with dd-source, but it is
generally useful, even to other repos. So perhaps it could be a
standalone command with a self extracting .bzl file, so you could just
have it in your path. But then it would need an installer.

But also, every rule developer will most likely use this as a starting
point to add a little custom formatting around their particular needs.
So, like The Pirate's Code, this is more guidelines than rules.