Migrate WorkloadFilter Util Selections (#40330)
### What does this PR do?

Migrates the workload filtering utility selection methods to be part of
the WorkloadFiltering component interface.

### Motivation

Efficiency gains to avoid re-calculating the filter selection. Note this
is different from compiling the filter, but rather just selecting the
filterIDs to be used in the Evaluation of inclusion/exclusion.

In the profiles below we can see that a significant amount of CPU time
is spent on simply _choosing_ which filters to be used. This is because
accessing values from the `config.Component` is expensive. Instead we
are storing the pre-computed constant selection within the
`workloadfilter.Component` and exposing it for clients to use.

<img width="330" height="267" alt="image"
src="https://github.com/user-attachments/assets/94090b7f-c2fc-48bc-82f4-25bcb1d78fb8"
/>
 
### Describe how you validated your changes (if not by through tests)

Unit tests to verify that the same selections are being made across each
client

Context: https://github.com/DataDog/datadog-agent/pull/39694 to move
away from using functions to component methods.