[CWS] Add eBPF support of prctl syscall (#40340)
This is just a template. Please delete all unneeded sections and add
others if appropriate.

### What does this PR do?
Add eBPF support of `prctl` syscall
Information available:
`prctl.option` : the option that is used for the syscall
`prctl.new_name` : if `option == PR_SET_NAME`, get the new_name of the
process
`prctl.is_name_truncated` : if `option == PR_SET_NAME`, indicates if the
retrieved name was truncated or not (there is also a metric tracking the
number of truncated names)

### Motivation
Prctl was not tracked before and can be used in different detection
rules increasing our coverage.

### Describe how you validated your changes (if not by through tests)
2 Functional tests : 
- one to test prctl with option `PR_SET_NAME`
- one to test prctl with option `PR_SET_NAME` and a truncated name (+
test the metric)


### Possible Drawbacks / Trade-offs

---------

Co-authored-by: domalessi <111786334+domalessi@users.noreply.github.com>