[macos test] Disable macos test in flake finder (#40545)
### What does this PR do?


Disable MacOS install script test in flake finder.
Flake finder is a pipeline that runs the test several time on nightlies.
For cost reasons and limit in the number of macos host we can have we
should disable the macos test in that pipeline

### Motivation

Make sure we do not go over the limit of dedicated host we have

### Describe how you validated your changes

### Additional Notes