fix: provide the correct version when creating the release PR (#40534)
### What does this PR do?

Provides the `notes.update_changelog` task with the correct version
number. The [solution
proposal](https://docs.google.com/document/d/1QrisX00J7Yf3swn-xgikHqtQKPvZxOe_mAd6H8lDb9o/edit?usp=sharing)
outlines why the current value is the correct one.

### Motivation

At the end of the release process, we run [this
command](https://datadoghq.atlassian.net/wiki/spaces/agent/pages/3649801154/Release+Coordinator+guide#Update-changelog)
to update the changelog.

The milestone associated with the generated PR is not the right one. For
instance, if we’re trying to update the changelog for 7.69.2, it will
have a 7.69.3 milestone, example
[here](https://github.com/DataDog/datadog-agent/pull/40024), instead of
7.69.2

The reason is because we use the value of the milestone from the
release.json file when we create the PR, but it was bumped as part of
[this
command](https://datadoghq.atlassian.net/wiki/spaces/agent/pages/3649801154/Release+Coordinator+guide#Prepare-final-release.json)
before we update the changelog

https://datadoghq.atlassian.net/browse/BARX-1165

### Describe how you validated your changes

### Additional Notes