[SBOM] implement spread refresher logic (#40379)
### What does this PR do?

This PR adds a new strategy for refreshing the SBOM of stored container
images. Instead of refreshing all images every 1h (which is what the
batch refresher is doing), this new spread refresher is refreshing 1/10
of all images every 1h/10 (6min). The 1/10 set of images are selected to
be the oldest images. This guarantees that all images will still be
updated in ~1h.

This new strategy is under a config flag (disabled by default) while
this undergoes further internal testing.

### Motivation

### Describe how you validated your changes (if not by through tests)

### Possible Drawbacks / Trade-offs