dyninst/decode: give interfaces an envelope
This may not be the right long term solution, but we need some mechanism
to resolve a conflict between our desire to flatten pointer chains and
our desire to display the concrete type within an interface box. Imagine
we have a scenario where there's a pointer to an interface. What should
the type be for that value? For the case of `*int`, we currently encode
it as having the type `"*int"` but the value will be that of an int. Now
imagine we have a `"*any"` and then inside that any is an `"int"`, what
should the type be?

After this change, it'll look like:

```json
{
	"type": "*any",
	"fields": {
		"data": {
			"type": "int",
			"value: "4"
		}
	}
}
```