Add support to take actions on Argo Rollouts objects in Autoscaling (#40570)
### What does this PR do?

Add support to take actions on `Rollout` objects in Autoscaling. There's
not a lot of changes as mostly the Cluster Agent supports any object,
_except_ for vertical, which what this PR addresses mostly.

### Motivation

Support Argo Rollouts

### Describe how you validated your changes

Currently you need custom code to "fake" feed the Autoscaling by
updating the `postProcess` function in
`pkg/clusteragent/autoscaling/workload/config_retriever_values.go` to
inject fake data for Argo Rollout.
Once the backend has proper support it can be easily validated
end-to-end by creating a `DPA` with a target to an Argo Rollout:
```
  kind: DatadogPodAutoscaler
  spec:
    owner: Local
    targetRef:
      apiVersion: argoproj.io/v1alpha1
      kind: Rollout
      name: rollout-canary
```

### Additional Notes

The `StrategicMergePatchType` is not supported for CRs, but for what we
are doing (updating a `map`), there's no difference between
`StrategicMergePatchType` and `MergePatchType (JSON Merge Patch)`