dyninst: remove with-depth-limits tests
These tests were poorly conceived. I wanted to make sure that
the tests still ran with a low depth limit, but I wasn't
validating much. The way to get the test to avoid having
expectations was to tell it it was in rewrite mode. The
downside of being in rewrite mode is we do not know how many
events to expect, so we just wait a while. This turns out to
be problematic and can cause flakes in production. We do exercise
a bunch of depth limits explicitly. For now, just remove this
bad subtest.