[backport] [netpath] Fix file descriptor handling on linux (#40639) (#40658)
This is a manual backport of #40639 to 7.71.x due to merge conflicts.

### What does this PR do?
This PR fixes an issue where datadog-traceroute was potentially
incorrectly handling file descriptors on linux. More details on the
datadog-traceroute PR: [\[linux\] Fix linuxSink file descriptor
handling](https://github.com/DataDog/datadog-traceroute/pull/24)

### Motivation
Errors from the current 7.71 RC

### Describe how you validated your changes
This was triggered by the finalizer of `*os.File` -- the datadog
traceroute library now has zero manual calls to unix.Close() on linux
which was the source of this issue. It purely uses os.File's handling to
close files