Fix OTLP testutil to avoid crashing when receiving sketches (#40628)
### What does this PR do?

This PR fixes the mock "logs" backend defined in
`comp/otelcol/otlp/testutil` so that it can receive requests to unknown
endpoints without crashing. At the moment, all unknown requests are
treated like requests to `/`, which results in attempting to parse them
as gzip-compressed JSON, crashing the test if that isn't the correct
format. By setting the path to `/{$}` instead, we limit this processing
to the `/` path.

### Motivation

One of our integration tests for the Datadog exporter [started
failing](https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/42513)
because it's exporting the OTel Collector's internal telemetry, which
since a recent PR now includes a histogram metric, meaning the exporter
makes a call to the sketch endpoint, which the mock backend doesn't have
a handler for, and which uses a very different format.