Add agent specific gcc toolchains (#40498)
### What does this PR do?
Introduces hermetic and downloadable gcc toolchains for Linux x86 and
aarch64 for bazel.

### Motivation
To ensure compatibility with older versions of `libc` we are relying on
cross-compilation toolchains built via [crosstool-ng
](https://github.com/crosstool-ng/crosstool-ng/). To take the next
logical step we want to let bazel handle toolchains. We are reusing
[gcc-toolchain](https://github.com/f0rmiga/gcc-toolchain) project that
represents a fully hermetic toolchain that, however, requires certain
adaptations to make it work with those created by `crosstool-ng`.

### Describe how you validated your changes
Used `toolchain_resolution_debug` flag and `--subcommands` to ensure we
are calling the tools from our hermetic toolchains.