Add version badge to README.md (#40027)
<!--
* Contributors are encouraged to read our
[CONTRIBUTING](/CONTRIBUTING.md) documentation.
* Both Contributor and Reviewer Checklists are available at
https://datadoghq.dev/datadog-agent/guidelines/contributing/#pull-requests.
* The pull request:
  * Should only fix one issue or add one feature at a time.
  * Must update the test suite for the relevant functionality.
  * Should pass all status checks before being reviewed or merged.
* Commit titles should be prefixed with general area of pull request's
change.
* Please fill the below sections if possible with relevant information
or links.
-->
### What does this PR do?

This PR adds a badge with the latest release number: 
<img width="397" height="161" alt="Screenshot 2025-08-19 at 2 11 13 PM"
src="https://github.com/user-attachments/assets/bf94c6b1-5185-4e4e-8def-f316e72a995c"
/>

### Motivation

I often struggle to find an immediate answer to the question: what is
the latest version of the agent.

I sometimes try to find this info by looking at the
[releases](https://github.com/DataDog/datadog-agent/releases) page, but
we often have a draft release on top. We also can have pre-releases
there.

### Describe how you validated your changes

This is a visual change to our documentation, should be validate by
inspecting rendered README.md.

### Possible Drawbacks / Trade-offs

If this is not the right place for this information, in case we want to
have this info somewhere else for consistency, this could become
duplicate, unnecessary information.

### Additional Notes
<!--
* Anything else we should know when reviewing?
* Include benchmarking information here whenever possible.
* Include info about alternatives that were considered and why the
proposed
  version was chosen.
-->