CONTINT-4685 - Add e2e tests for host-tags presence and values
Add a new generic function that reads host-tags from `fakeintake`, then
validates the received tags against the expected tags for the tests for
that env.

Each env runs the test with its own expected tags and values.
Some env runs with different host-type (windows and linux) and so the
expected list of tags is different, we must use the optional tags for
the the extra tags that are only here on some host types.

Signed-off-by: Alexandre Lavigne <alexandre.lavigne@datadoghq.com>